requirements.txt
alive-progress==1.6.2
aniso8601==9.0.1
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
h11==0.12.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
prompt-toolkit==1.0.14
psutil==5.8.0
pycparser==2.20
pydantic==1.8.2
pyfiglet==0.8.post1
Pygments==2.9.0
PyInquirer==1.0.3
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
regex==2021.7.6
requests==2.26.0
rq==1.9.0
six==1.16.0
starlette==0.14.2
tabulate==0.8.9
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
